Removed universality of Page level class toggle. (Might be added later.)
