+++ src/components/ModalInMobile/ModalInMobile.example.js
@@ -1,23 +1,19 @@
 /* eslint-disable no-console, import/prefer-default-export */
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
 import { Button } from '../../components';
 import ModalInMobile from './ModalInMobile';
 import css from './ModalInMobile.example.css';
+
+const togglePageClassNames = (className, addClass = true) => {
+  // We are just checking the value for now
+  console.log('Toggling ModalInMobile - currently:', className, addClass);
+}
 
 class ModalInMobileWrapper extends Component {
   constructor(props) {
     super(props);
     this.state = { isOpen: false };
     this.handleOpen = this.handleOpen.bind(this);
-  }
-
-  getChildContext() {
-    return {
-      togglePageClassNames: (className, addClass = true) => {
-        // We are just checking the value for now
-        console.log('Toggling ModalInMobile - currently:', className, addClass);
-      },
-    };
   }
 
   handleOpen() {
@@ -28,7 +24,10 @@ class ModalInMobileWrapper extends Component {
     return (
       <div>
         Wrapper text before ModalInMobile
-        <ModalInMobile {...this.props} isModalOpenOnMobile={this.state.isOpen}>
+        <ModalInMobile
+          {...this.props}
+          isModalOpenOnMobile={this.state.isOpen}
+          togglePageClassNames={togglePageClassNames}>
           Some content inside ModalInMobile component
         </ModalInMobile>
         <Button onClick={this.handleOpen} className={css.visibleOnMobileLayout}>Open</Button>
@@ -36,8 +35,6 @@ class ModalInMobileWrapper extends Component {
     );
   }
 }
-
-ModalInMobileWrapper.childContextTypes = { togglePageClassNames: PropTypes.func.isRequired };
 
 export const Empty = {
   component: ModalInMobileWrapper,
